refactor projects
