+++ tests/Feature/ExampleTest.php
@@ -14,7 +14,7 @@ class ExampleTest extends TestCase
      */
     public function testBasicTest()
     {
-        $response = $this->get('/');
+        $response = $this->get('/api/projects');
 
         $response->assertStatus(200);
     }
